% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barrier}
\alias{barrier}
\title{Map of the barrier}
\format{
An \code{sf} object from original shapefiles
}
\usage{
barrier
}
\description{
A map of the barrier for (no)prediction
}
\keyword{datasets}
