% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_points.R
\name{check_points}
\alias{check_points}
\title{Check for points on a polygon}
\usage{
check_points(points, polygon)
}
\arguments{
\item{points}{A sf object with points (usually of class
\code{sfc_POINT})}

\item{polygon}{A sf object with a polygon}
}
\value{
A list with the number of points on the polygon and the
coordinates of the points on the polygon. If there are no points on
the polygon, only the number of points is returned.
}
\description{
This function checks if there are points on a polygon.
}
\details{
This function checks if there are points on a polygon. If
there are points on the polygon, it returns the number of points and
the coordinates of the points. If there are no points on the
polygon, it returns only the number of points.
}
\author{
Fernando Mayer
}
