% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indices.R
\name{create_indices}
\alias{create_indices}
\title{Create Indices for the Spatial Field and Temporal Components}
\usage{
create_indices(time_indices, mesh)
}
\arguments{
\item{time_indices}{A vector with the time indices.}

\item{mesh}{A mesh object.}
}
\value{
A list with the indices for the spatial field and temporal
components.
}
\description{
This function creates indices for the spatial field and
temporal components.
}
\details{
The function creates indices for the spatial field and
temporal components. The function is used to create the indices for
the INLA model.
}
\author{
Fernando Mayer
}
