% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_layer.R
\name{crop_layer}
\alias{crop_layer}
\title{Crop a layer}
\usage{
crop_layer(shape, xmin, ymin, xmax, ymax, crs = NULL)
}
\arguments{
\item{shape}{An \code{sf} object}

\item{xmin}{The minimum x-coordinate}

\item{ymin}{The minimum y-coordinate}

\item{xmax}{The maximum x-coordinate}

\item{ymax}{The maximum y-coordinate}

\item{crs}{The coordinate reference system of the output layer. If
\code{NULL}, the CRS of the input layer is used.}
}
\value{
An \code{sf} object
}
\description{
Crop a layer to a specified extent using an \code{sf}
object as the extent.
}
\details{
This function crops a layer to a specified extent using an
\code{sf} object as the extent. The function takes an \code{sf}
object and the minimum and maximum x and y coordinates of the extent
to crop the layer. The function returns an \code{sf} object with the
same projection as the input layer.
}
\author{
Fernando Mayer
}
