% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_imports.R
\name{extract_imports}
\alias{extract_imports}
\title{Extract imported functions and packages}
\usage{
extract_imports(pkg_path = ".")
}
\arguments{
\item{pkg_path}{The path to the package directory. Default is the
current working directory.}
}
\value{
A list of data frames, one for each R file, with the
imported functions and packages.
}
\description{
Extract imported functions and packages from a
package's R files that are documented with roxygen2-style comments.
}
\details{
This function reads the R files in the R/ directory of a
package and extracts the imported functions and packages that are
documented with roxygen2-style comments. The function returns a list
of data frames, one for each R file, with the imported functions and
packages. The packages are extracted from lines that start with the
text "@import" and the functions are extracted from lines that start
with the text "@importFrom".
}
\author{
Fernando Mayer
}
