% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plot.R
\name{field_plot}
\alias{field_plot}
\title{Animate the model output for a given field and barrier}
\usage{
field_plot(data, barrier, grid, field, crs)
}
\arguments{
\item{data}{A sf object that represents the data points to be
plotted on the map.}

\item{barrier}{A sf object that represents the barrier.}

\item{grid}{A list of x and y values that represent the grid points.}

\item{field}{A list of matrices, each matrix is a 2D array of the
field values at a given year.}

\item{crs}{The Coordinate Reference System of the data points}
}
\value{
A gif file that represents the animation of the model output.
}
\description{
This function animates the model output for a given
field and barrier. The field is a list of matrices, each matrix is a
2D array of the field values at a given year. The barrier is a sf
object that represents the barrier. The grid is a list of x and y
values that represent the grid points. The data is a sf object that
is used to plot the points. The name is the name of the output file.
}
\details{
The function uses the gganimate package to animate the
model output.
}
\author{
Fernando Mayer
}
