% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_fields.R
\name{fields}
\alias{fields}
\alias{mfield}
\alias{sfield}
\title{Field functions}
\usage{
mfield(m, index, projgrid, crs)

sfield(m, index, projgrid, crs)
}
\arguments{
\item{m}{INLA model}

\item{index}{Index of the time variable in the data}

\item{projgrid}{Projected grid to project the field to the original
grid of the data (not the mesh)}

\item{crs}{The Coordinate Reference System of the data points}
}
\value{
A list with the mean field at each time point
}
\description{
Functions to extract and plot fields from INLA models
with spatial and temporal components.
}
\details{
The functions \code{mfield} and \code{sfield} extract the
mean and standard deviation fields from an INLA model with spatial
and temporal components. The function \code{plot_field} plots the
field at a given time.
}
\author{
Fernando Mayer
}
