% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev.R
\name{fix_lengths}
\alias{fix_lengths}
\title{Fix lengths of variables}
\usage{
fix_lengths(...)
}
\arguments{
\item{...}{A list of variable names.}
}
\value{
The function returns 0 invisibly.
}
\description{
Fix lengths of variables by repeating them to the
length of the longest variable.
}
\details{
This function takes a list of variable names and repeats
them to the length of the longest variable. If the length of the
longest variable is not a multiple of the length of the other
variables, an error is thrown. The variables are modified in place.
}
\author{
Daniela Castro-Camilo
}
