% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_residuals.R
\name{get_fit_res}
\alias{get_fit_res}
\title{Get fit and residuals from INLA model}
\usage{
get_fit_res(mod, y, idat, fun = NULL, std = TRUE)
}
\arguments{
\item{mod}{An INLA model object.}

\item{y}{A numeric vector of observed values.}

\item{idat}{A numeric vector of indices.}

\item{fun}{A function to extract the fitted values and residuals.}

\item{std}{A logical value indicating whether to standardize the
residuals. Default is \code{TRUE}.}
}
\value{
A data frame with the fitted values and residuals.
}
\description{
Get fit and residuals from INLA model for a given
dataset and model.
}
\details{
This function extracts the fitted values and residuals from
an INLA model for a given dataset. The function can also be used to
extract the fitted values and residuals from a given function.
}
\author{
Fernando Mayer
}
