% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpd.R
\name{get_hpd}
\alias{get_hpd}
\title{Get HPD intervals for fixed effects from INLA model}
\usage{
get_hpd(mod, p = 0.95)
}
\arguments{
\item{mod}{INLA model object}

\item{p}{Probability level for the HPD interval}
}
\value{
A data frame with the HPD intervals for each fixed effect in
the model object.
}
\description{
Get HPD intervals for fixed effects from INLA model
using the inla.hpdmarginal function from the INLA package.
}
\details{
The function takes an INLA model object and a probability
level p (default is 0.95) and returns a data frame with the HPD
intervals for each fixed effect. The intervals are calculated using
the inla.hpdmarginal function from the INLA package.
}
\author{
Fernando Mayer
}
