% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_effect.R
\name{get_spatial_effect}
\alias{get_spatial_effect}
\title{Get spatial effect from an INLA model}
\usage{
get_spatial_effect(mod, stk_idx)
}
\arguments{
\item{mod}{An INLA model object.}

\item{stk_idx}{Stacked indexes of the observations.}
}
\value{
A data frame with three columns corresponding to the three
quantiles of the spatial effect.
}
\description{
Get spatial effect from an INLA model and return it as
a data frame.
}
\details{
This function extracts the spatial effect from an INLA
model and returns it as a data frame. The data frame has three
columns corresponding to the three quantiles of the spatial effect.
}
\author{
Fernando Mayer
}
