% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stack_index.R
\name{get_stack_index}
\alias{get_stack_index}
\title{Get stacked indexes from INLA object}
\usage{
get_stack_index(stk, type = "est")
}
\arguments{
\item{stk}{INLA object to extract indexes from (output from
\code{inla}).}

\item{type}{Type of indexes to extract (est, obs, pred).}
}
\value{
A data frame with the indexes of the stacked data.
}
\description{
Get stacked indexes from INLA object for a given type
(est, obs, pred) and return them as a data frame.
}
\details{
This function is a wrapper for the \code{inla.stack.index}
function from the \code{INLA} package. It returns the indexes of the
stacked data for a given type (est, obs, pred) as a data frame.
}
\author{
Fernando Mayer
}
