% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev.R
\name{gev_internals}
\alias{gev_internals}
\alias{new_to_old}
\alias{old_to_new}
\alias{pbgev}
\alias{qbgev}
\alias{rbgev}
\alias{dbgev}
\alias{dbgev2}
\alias{return_level_bgev}
\alias{return_level_bgev2}
\alias{dbgev_mixing}
\alias{qbgev_mixing}
\alias{get_gumbel_par}
\alias{pgev}
\alias{qgev}
\alias{rgev}
\alias{dgev}
\alias{dgev2}
\alias{return_level_gev}
\alias{nllik_gev}
\alias{nllik_bgev}
\alias{nllik_gevx}
\alias{nllik_bgevx}
\alias{nllik_gev2}
\alias{nllik_bgev2}
\alias{map_tail}
\title{Internal GEV helpers}
\usage{
new_to_old(par, alpha = 0.5, beta = 0.5)

old_to_new(par, alpha = 0.5, beta = 0.5)

pbgev(x, mu, sigma, xi, p_a = 0.1, p_b = 0.2, s = 5)

qbgev(p, mu, sigma, xi, p_a = 0.1, p_b = 0.2, s = 5)

rbgev(n, mu, sigma, xi, p_a = 0.1, p_b = 0.2, s = 5)

dbgev(x, mu, sigma, xi, p_a = 0.1, p_b = 0.2, s = 5, log = FALSE)

dbgev2(
  x,
  q,
  sb,
  xi,
  alpha = 0.5,
  beta = 0.5,
  p_a = 0.1,
  p_b = 0.2,
  s = 5,
  log = FALSE
)

return_level_bgev(period, mu, sigma, xi, p_a = 0.1, p_b = 0.2, s = 5)

return_level_bgev2(
  period,
  q,
  sb,
  xi,
  alpha = 0.5,
  beta = 0.5,
  p_a = 0.1,
  p_b = 0.2,
  s = 5
)

dbgev_mixing(x, mu, sigma, xi, p_a = 0.1, p_b = 0.2, s = 5, log = FALSE)

qbgev_mixing(
  p,
  mu,
  sigma,
  xi,
  p_a = 0.1,
  p_b = 0.2,
  s = 5,
  lower = 0,
  upper = 100
)

get_gumbel_par(mu, sigma, xi, p_a = 0.1, p_b = 0.2)

pgev(x, mu, sigma, xi)

qgev(p, mu, sigma, xi)

rgev(n, mu, sigma, xi)

dgev(x, mu, sigma, xi, log = FALSE)

dgev2(x, q, sb, xi, alpha = 0.5, beta = 0.5, log = FALSE)

return_level_gev(period, mu, sigma, xi)

nllik_gev(par, x, log = TRUE)

nllik_bgev(par, x, p_a, p_b, s, log = TRUE)

nllik_gevx(par, x, w, log = TRUE)

nllik_bgevx(par, x, w, p_a, p_b, s, log = TRUE)

nllik_gev2(par, x, alpha = 0.5, beta = 0.5, log = TRUE)

nllik_bgev2(par, x, alpha = 0.5, beta = 0.5, p_a, p_b, s, log = TRUE)

map_tail(x, interval, inverse = FALSE)
}
\description{
Random generation, density, distribution and related functions for the GEV
distribution. These are exported primarily for use by other packages and
advanced users, and are not part of the stable user-facing API.
}
\keyword{internal}
