% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_all}
\alias{map_all}
\title{Map of Republic of Ireland and Great Britain}
\format{
An \code{sf} object from \code{rnaturalearth}
}
\usage{
map_all
}
\description{
A map of Republic of Ireland and Great Britain
}
\keyword{datasets}
