% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_irl}
\alias{map_irl}
\title{Map of Republic of Ireland and Northern Ireland}
\format{
An \code{sf} object from \code{rnaturalearth}
}
\usage{
map_irl
}
\description{
A map of Republic of Ireland and Northern Ireland
}
\keyword{datasets}
