% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_roi}
\alias{map_roi}
\title{Map of Republic of Ireland}
\format{
An \code{sf} object from \code{rnaturalearth}
}
\usage{
map_roi
}
\description{
A map of Republic of Ireland
}
\keyword{datasets}
