% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_st_gev.R
\name{marginals_st_gev}
\alias{marginals_st_gev}
\title{Marginal posterior distributions of the parameters}
\usage{
marginals_st_gev(mod)
}
\arguments{
\item{mod}{INLA model object}
}
\value{
A list of tibbles, one for each parameter.
}
\description{
This function extracts the marginal posterior
distributions of the parameters from an INLA model object.
}
\details{
The function extracts the marginal posterior distributions
of the parameters from an INLA model object. The function returns a
list of tibbles, one for each parameter. Each tibble contains the
marginal posterior distribution of the parameter.
}
\author{
Fernando Mayer
}
