% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary.R
\name{mod_summary}
\alias{mod_summary}
\title{General summary of a model diagnostics}
\usage{
mod_summary(mod, y, idx, fun = sum, na.rm = TRUE, KIS = TRUE)
}
\arguments{
\item{mod}{An object of class \code{inla}}

\item{y}{The (observed) response variable}

\item{idx}{The index from INLA stack}

\item{fun}{Function used to summarise the values}

\item{na.rm}{Remove NAs (defaults to \code{TRUE})}

\item{KIS}{Keep It Simple: keep extrictly the main measures only (don't show
number of parameters, fails and CPU time)}
}
\value{
A named vector with the main diagnostics of the model.
}
\description{
This function returns a data frame with the main diagnostics of
an INLA model (DIC, WAIC, lCPO, etc).
}
\details{
NEEDS UPDATE. The function returns a data frame with the following
columns: DIC, NP.DIC, WAIC, NP.WAIC, CPO, FAIL, NFAIL, CPU. The first
four columns are the main diagnostics of the model, the next two are the
cross-validation diagnostics, and the last one is the CPU time used by
the model.
}
\author{
Fernando Mayer
}
