% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_st_gev.R
\name{plot.res_marg}
\alias{plot.res_marg}
\title{Plot marginal posterior distributions of the parameters}
\usage{
\method{plot}{res_marg}(x, ...)
}
\arguments{
\item{x}{A list of tibbles, one for each parameter.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\value{
A plot with the marginal posterior distributions of the
parameters.
}
\description{
This function plots the marginal posterior
distributions of the parameters.
}
\details{
The function plots the marginal posterior distributions of
the parameters. The function returns a plot with the marginal
posterior distributions of the parameters.
}
\author{
Fernando Mayer
}
