% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_on_mesh.R
\name{plot.tri_on_mesh}
\alias{plot.tri_on_mesh}
\title{Plot triangles on a mesh}
\usage{
\method{plot}{tri_on_mesh}(x, mesh, all = FALSE, ...)
}
\arguments{
\item{x}{A \code{tri_on_mesh} object}

\item{mesh}{An \code{inla.mesh} object}

\item{all}{A logical indicating if all triangles should be plotted
or only the triangles that are within the polygon}

\item{...}{Additional arguments to be passed to
\code{\link[ggplot2]{geom_point}} function}
}
\value{
A ggplot object
}
\description{
This function plots the triangles on a mesh that are
within a polygon.
}
\author{
Fernando Mayer
}
