% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_pred.R
\name{poly_pred}
\alias{poly_pred}
\title{Create a polygon for the spatial prediction}
\usage{
poly_pred(shape, xmin, ymin, xmax, ymax, crs = NULL)
}
\arguments{
\item{shape}{An \code{sf} object}

\item{xmin}{The minimum x-coordinate}

\item{ymin}{The minimum y-coordinate}

\item{xmax}{The maximum x-coordinate}

\item{ymax}{The maximum y-coordinate}

\item{crs}{The coordinate reference system of the output layer. If
\code{NULL}, the CRS of the input layer is used.}
}
\value{
A polygon with the area to predict
}
\description{
This function creates a polygon with the area to
predict based on the input shapefile and the coordinates of the
area.
}
\details{
The function receives a shapefile and the coordinates of
the area to predict. The function creates a polygon with the area to
predict based on the input shapefile and the coordinates of the area.
}
\author{
Fernando Mayer
}
