% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_all}
\alias{shp_all}
\title{SHP of Republic of Ireland and Great Britain}
\format{
An \code{sf} object from original shapefiles
}
\usage{
shp_all
}
\description{
A map of Republic of Ireland and Great Britain (shapefile version)
}
\keyword{datasets}
