% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_gbr}
\alias{shp_gbr}
\title{SHP of Great Britain}
\format{
An \code{sf} object from original shapefiles
}
\usage{
shp_gbr
}
\description{
A map of Great Britain (shapefile version)
}
\keyword{datasets}
