% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_st_gev.R
\name{summary.res_marg}
\alias{summary.res_marg}
\title{Summary of the marginal posterior distributions of the
parameters}
\usage{
\method{summary}{res_marg}(object, round = NULL, ...)
}
\arguments{
\item{object}{A list of tibbles, one for each parameter.}

\item{round}{Number of decimal places to round the results.}

\item{...}{Additional arguments to be passed to the summary function.}
}
\value{
A table with the mean, median, mode, standard deviation,
2.5\% quantile, and 97.5\% quantile of the marginal posterior
distributions of the parameters.
}
\description{
This function summarizes the marginal posterior
distributions of the parameters.
}
\details{
The function summarizes the marginal posterior
distributions of the parameters. The function returns a table with
the mean, median, mode, standard deviation, 2.5\% quantile, and 97.5\%
quantile of the marginal posterior distributions of the parameters.
}
\author{
Fernando Mayer
}
