% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_on_mesh.R
\name{tri_on_mesh}
\alias{tri_on_mesh}
\title{Identify triangles on a mesh that are within a polygon}
\usage{
tri_on_mesh(mesh, polygon, inverse = FALSE)
}
\arguments{
\item{mesh}{An \code{inla.mesh} object}

\item{polygon}{A \code{sf} object}
}
\value{
A vector of integers indicating the triangles that are
within the polygon and a \code{tibble} with the centroids of the
triangles that are within the polygon.
}
\description{
This function identifies the triangles on a mesh that
are within a polygon.
}
\details{
This function uses the \code{fmesher} package to identify
the triangles on a mesh that are within a polygon. The function
returns a vector of integers indicating the triangles that are
within the polygon and a \code{tibble} with the centroids of the
triangles that are within the polygon.
}
\author{
Fernando Mayer
}
