% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stExtremes-package.R
\docType{package}
\name{stExtremes-package}
\alias{stExtremes}
\alias{stExtremes-package}
\title{stExtremes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Fernando Mayer \email{fernando.mayer@mu.ie} (\href{https://orcid.org/0000-0001-5115-338X}{ORCID})

Authors:
\itemize{
  \item Niamh Cahill \email{niamh.cahill@mu.ie} (\href{https://orcid.org/0000-0003-3086-550X}{ORCID})
}

}
\keyword{internal}
